joined
joined try concession is misleading, expecting a throw if model does not exists. yet works flawlessly event model not in query. this separation between layout and joined keeps existing functionality and extends it by desired behaviour to disrupt execution once result does not meet expectation. layout is a plain copy of the function joined and its intend is to retrieve the layout of any model as a fluent object. ion the otherhand joined should retrieve the model from a query and throw once this not happens.